requirements_hackathon.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
transformers>=4.35.0

# 3D processing
open3d>=0.18.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Existing dependencies
scipy>=1.11.0
matplotlib>=3.7.0